refactor: create LottoReport class
